requirements.txt
asttokens==3.0.0
click==8.3.0
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
filelock==3.19.1
fonttools==4.60.1
fsspec==2025.9.0
ghp-import==2.1.0
iniconfig==2.1.0
ipykernel==7.0.1
ipython==9.6.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.2
jupyter-client==8.6.3
jupyter-core==5.9.1
kiwisolver==1.4.9
loguru==0.7.3
markdown==3.9
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.7
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.4
packaging==25.0
pandas==2.3.3
parso==0.8.5
pathspec==0.12.1
pillow==11.3.0
pip==25.2
platformdirs==4.5.0
pluggy==1.6.0
prompt-toolkit==3.0.52
psutil==7.1.0
pure-eval==0.2.3
pygments==2.19.2
pyparsing==3.2.5
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyyaml==6.0.3
pyyaml-env-tag==1.1
pyzmq==27.1.0
rich==14.2.0
-e file:///C:/Users/dgmar/Isa/rps-image-classification
ruff==0.14.0
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
setuptools==70.2.0
shellingham==1.5.4
six==1.17.0
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.6.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.19.2
typing-extensions==4.15.0
tzdata==2025.2
watchdog==6.0.0
wcwidth==0.2.14
win32-setctime==1.2.0
